Decouple auth components
